+++ models/users.js
@@ -17,6 +17,8 @@ const { _U_ }               = require('../controllers/utilities/');
 //  SCHEMA
 // ==============================
 const UserSchema            = new mongoose.Schema({
+    active                  : { type: String, default: false },
+    roles                   : { type: String, default: 'admin' },
     username                : { type: String, unique: true, lowercase: true },
     email                   : { type: String, unique: true, required: true,
         validate: {
@@ -25,11 +27,18 @@ const UserSchema            = new mongoose.Schema({
             message         : 'Invalid email address',
         }},
     nickname                : { type: String },
-    firstName               : { type: String, required: true },
-    lastName                : { type: String, required: true },
-    picture                 : { type: String, required: true, default: '' },
-    roles                   : { type: String, default: 'admin' },
-    active                  : { type: String, default: false },
+    picture                 : { type: String, required: true, default: '' },    // todo: added a validator
+    info: {
+        firstName           : { type: String, required: true },
+        lastName            : { type: String, required: true },
+        gender              : { type: String, required: true, enum: ['Male', 'Female', 'NA'], default: 'NA' },
+        residence           : { type: String },
+        timezone            : { type: String },
+        birthday            : { type: Date },
+    },
+    time: {
+        _lastTimeSignIn     : { type: Date },
+    },
 }, {
     timestamps              : { createdAt: 'time._registered', updatedAt: 'time._updated' },
     versionKey              : false,
@@ -42,15 +51,22 @@ const UserSchema            = new mongoose.Schema({
 // ==============================
 // nickname assignment (pre-hook)
 UserSchema.pre('save', function () {
-    if (this.nickname === undefined) this.nickname = `${this.firstName} ${this.lastName}`;
+    if (this.nickname === undefined) this.nickname = `${this.info.firstName} ${this.info.lastName}`;
 });
 
+// methods for the document
+UserSchema.methods.UpdateSignInLog = function () {
+    return mongoose.connection.db.collection('users').update(
+        { _id: this._id },
+        { $set: { 'time._lastTimeSignIn': new Date(Date.now()) }}
+    );
+};
 
 // methods from third-party plugin (object method)
 UserSchema.plugin(passportLocalMongoose, {
     usernameField: 'email',
     usernameQueryFields: ['username'],
-    selectFields: ['_id', 'email', 'username', 'nickname', 'picture'],
+    selectFields: ['_id', 'email', 'nickname', 'picture', 'info', 'time'],
 });
 
 //// rewrite plugin methods as promises
@@ -68,7 +84,6 @@ const _changePassword = UserSchema.methods.changePassword;
 UserSchema.methods.changePassword = function (old_PW, new_PW, next) {
     return _U_.schema.promisify(_changePassword, arguments, this);
 };
-
 
 
 // exports